Gestion réglemnts joueurs

+++ jquery/jQuery-File-Upload/server/php/UploadHandler.php
@@ -327,7 +327,8 @@ class UploadHandler
     }
 
     function get_config_bytes($val) {
-        $val = trim($val);
+        $val = (int)trim($val);
+        
         $last = strtolower($val[strlen($val)-1]);
         switch($last) {
             case 'g':
